requirements.txt
streamlit
pandas
qrcode
Pillow
reportlab
